print package versions at startup
