+++ src/hdr-exporter.js
